Recovered from Surgery, Coding Again
Hit a Roadblock with Displaying messages so I began implementation of a friend system to the database, Users can now add a friend using their email. Reworked the backend functionality of all the screens to make it easier to code and installed the global variable UserID to each instance allowing easier transition between pages.